pyproject.toml
[project]
name = "sora-videogen"
version = "0.1.0"
description = "Generate narrated video summaries from URLs using OpenAI GPT-5-mini and Sora 2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "openai>=2.29.0,<3",
    "azure-identity>=1.20.0",
    "azure-cognitiveservices-speech>=1.48.2",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.18",
    "jinja2>=3.1.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
